Improved Deployment Page - Improved workload editor - Improved Env Cards & Status Indicators
