+++ plugins/openchoreo/src/components/Environments/EnvCard/index.ts
@@ -1,0 +1,1 @@
+export { EnvCard } from './EnvCard';
