+++ plugins/openchoreo/src/components/Environments/Environments.tsx
@@ -3,29 +3,22 @@ import { useCallback, useEffect, useState } from 'react';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { Content, Page } from '@backstage/core-components';
 import {
-  Grid,
   Card,
   CardContent,
   Typography,
   Box,
-  Button,
-  IconButton,
+  Collapse,
+  useTheme,
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import FileCopyIcon from '@material-ui/icons/FileCopy';
-import AccessTimeIcon from '@material-ui/icons/AccessTime';
+
 
 import {
   discoveryApiRef,
   identityApiRef,
   useApi,
 } from '@backstage/core-plugin-api';
-import {
-  fetchEnvironmentInfo,
-  promoteToEnvironment,
-  updateComponentBinding,
-} from '../../api/environments';
-import { formatRelativeTime } from '../../utils/timeUtils';
+import { fetchEnvironmentInfo } from '../../api/environments';
 
 interface EndpointInfo {
   name: string;
@@ -34,7 +27,9 @@ interface EndpointInfo {
   visibility: 'project' | 'organization' | 'public';
 }
 import { Workload } from './Workload/Workload';
-import Refresh from '@material-ui/icons/Refresh';
+import { EnvCard } from './EnvCard';
+import { useTimerEffect } from '../../hooks/timerEffect';
+import { Alert } from '@material-ui/lab';
 
 const useStyles = makeStyles(theme => ({
   notificationBox: {
@@ -59,50 +54,8 @@ const useStyles = makeStyles(theme => ({
     padding: theme.spacing(1),
     borderRadius: theme.shape.borderRadius,
   },
-  deploymentStatusBox: {
-    padding: theme.spacing(1),
-    borderRadius: theme.shape.borderRadius,
-    marginTop: theme.spacing(2),
-  },
-  successStatus: {
-    backgroundColor: theme.palette.success.light,
-    color: theme.palette.success.dark,
-  },
-  errorStatus: {
-    backgroundColor: theme.palette.error.light,
-    color: theme.palette.error.dark,
-  },
-  warningStatus: {
-    backgroundColor: theme.palette.warning.light,
-    color: theme.palette.warning.dark,
-  },
-  defaultStatus: {
-    backgroundColor: theme.palette.background.paper,
-    color: theme.palette.text.primary,
-  },
-  imageContainer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(1.5),
-    borderRadius: theme.spacing(3),
-    border: `1px solid ${theme.palette.divider}`,
-    boxShadow: theme.shadows[2],
-    marginTop: theme.spacing(1),
-  },
-  endpointLink: {
-    color: theme.palette.primary.main,
-    textDecoration: 'underline',
-    display: 'block',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-    fontSize: '0.875rem',
-  },
-  timeIcon: {
-    fontSize: '1rem',
-    color: theme.palette.text.secondary,
-    marginLeft: theme.spacing(1),
-    marginRight: theme.spacing(1),
-  },
+
+
 }));
 
 interface Environment {
@@ -129,13 +82,14 @@ export const Environments = () => {
   const [loading, setLoading] = useState(true);
   const [promotingTo, setPromotingTo] = useState<string | null>(null);
   const [updatingBinding, setUpdatingBinding] = useState<string | null>(null);
+  const [isWorkloadEditorOpen, setIsWorkloadEditorOpen] = useState(false);
   const [notification, setNotification] = useState<{
     message: string;
     type: 'success' | 'error';
   } | null>(null);
   const discovery = useApi(discoveryApiRef);
   const identityApi = useApi(identityApiRef);
-
+  const theme = useTheme();
   const fetchEnvironmentsData = useCallback(async () => {
     try {
       setLoading(true);
@@ -159,21 +113,7 @@ export const Environments = () => {
     env => env.deployment.status === 'pending',
   );
 
-  useEffect(() => {
-    let intervalId: NodeJS.Timeout;
-
-    if (isPending) {
-      intervalId = setInterval(() => {
-        fetchEnvironmentsData();
-      }, 10000); // 10 seconds
-    }
-
-    return () => {
-      if (intervalId) {
-        clearInterval(intervalId);
-      }
-    };
-  }, [isPending, fetchEnvironmentsData]);
+  useTimerEffect(fetchEnvironmentsData, isWorkloadEditorOpen || !isPending ? 0 : 10000, [isPending, fetchEnvironmentsData]);
 
   if (loading && !isPending) {
     return (
@@ -193,425 +133,62 @@ export const Environments = () => {
   }
 
   return (
-    <Page themeId="tool">
-      <Content>
-        {notification && (
-          <Box
-            className={`${classes.notificationBox} ${
-              notification.type === 'success'
-                ? classes.successNotification
-                : classes.errorNotification
-            }`}
-          >
-            <Typography variant="body2" style={{ fontWeight: 'bold' }}>
-              {notification.type === 'success' ? '✓ ' : '✗ '}
-              {notification.message}
-            </Typography>
-          </Box>
-        )}
-        <Grid container spacing={3}>
-          <Grid item xs={12} md={3}>
-            <Card>
-              {/* Make this card color different from the others */}
-              <Box className={classes.setupCard}>
-                <CardContent>
-                  <Typography variant="h6" component="h4">
-                    Set up
-                  </Typography>
+    <Box height="100%" p={2}>
+      <Collapse in={!!notification} mountOnEnter unmountOnExit>
+        <Alert severity={notification?.type} title={notification?.type === 'success' ? '✓ ' : '✗ '} style={{ marginBottom: '10px' }}>
+          {notification?.message}
+        </Alert>
+      </Collapse>
+      <Box display="flex" flexDirection="row" gridGap={16} py={1} overflow="auto" flexGrow={1} width="100%">
+        <Box minWidth={theme.spacing(40)} width={theme.spacing(40)} key="setup">
+          <Card>
+            {/* Make this card color different from the others */}
+            <Box className={classes.setupCard}>
+              <CardContent>
+                <Typography variant="h6" component="h4">
+                  Set up
+                </Typography>
 
-                  <Box
-                    borderBottom={1}
-                    borderColor="divider"
-                    marginBottom={2}
-                    marginTop={1}
+                <Box
+                  borderBottom={1}
+                  borderColor="divider"
+                  marginBottom={2}
+                  marginTop={1}
+                />
+                <Typography color="textSecondary">
+                  View and manage deployment environments
+                </Typography>
+                {isWorkloadEditorSupported && !loading && (
+                  <Workload
+                    isOpen={isWorkloadEditorOpen}
+                    onOpenChange={setIsWorkloadEditorOpen}
+                    onDeployed={fetchEnvironmentsData}
+                    isWorking={isPending}
                   />
-                  <Typography color="textSecondary">
-                    View and manage deployment environments
-                  </Typography>
-                  {isWorkloadEditorSupported && !loading && (
-                    <Workload
-                      onDeployed={fetchEnvironmentsData}
-                      isWorking={isPending}
-                    />
-                  )}
-                </CardContent>
-              </Box>
-            </Card>
-          </Grid>
-          {environments.map(env => (
-            <Grid key={env.name} item xs={12} md={3}>
-              <Card>
-                <CardContent>
-                  <Box
-                    display="flex"
-                    alignItems="center"
-                    justifyContent="space-between"
-                  >
-                    <Typography variant="h6" component="h4">
-                      {env.name}
-                    </Typography>
-                    <IconButton onClick={() => fetchEnvironmentsData()}>
-                      <Refresh
-                        fontSize="inherit"
-                        style={{ fontSize: '18px' }}
-                      />
-                    </IconButton>
-                  </Box>
-                  {/* add a line in the ui */}
-                  <Box
-                    borderBottom={1}
-                    borderColor="divider"
-                    marginBottom={2}
-                    marginTop={1}
-                  />
-                  {env.deployment.lastDeployed && (
-                    <Box display="flex" alignItems="center">
-                      <Typography variant="body2" color="textSecondary">
-                        Deployed
-                      </Typography>
-                      <AccessTimeIcon className={classes.timeIcon} />
-                      <Typography variant="body2" color="textSecondary">
-                        {formatRelativeTime(env.deployment.lastDeployed)}
-                      </Typography>
-                    </Box>
-                  )}
-                  <Box
-                    display="flex"
-                    alignItems="center"
-                    className={`${classes.deploymentStatusBox} ${
-                      env.deployment.status === 'success'
-                        ? classes.successStatus
-                        : env.deployment.status === 'failed'
-                        ? classes.errorStatus
-                        : env.deployment.status === 'pending'
-                        ? classes.warningStatus
-                        : env.deployment.status === 'suspended'
-                        ? classes.warningStatus
-                        : classes.defaultStatus
-                    }`}
-                  >
-                    <Typography variant="body2" color="textSecondary">
-                      Deployment Status:{' '}
-                      <span
-                        style={{
-                          fontWeight:
-                            env.deployment.status === 'success'
-                              ? 'bold'
-                              : 'normal',
-                        }}
-                      >
-                        {env.deployment.status === 'success'
-                          ? 'Active'
-                          : env.deployment.status === 'pending'
-                          ? 'Pending'
-                          : env.deployment.status === 'not-deployed'
-                          ? 'Not Deployed'
-                          : env.deployment.status === 'suspended'
-                          ? 'Suspended'
-                          : 'Failed'}
-                      </span>
-                    </Typography>
-                  </Box>
-                  {env.deployment.statusMessage && (
-                    <Box mt={1}>
-                      <Typography variant="caption" color="textSecondary">
-                        {env.deployment.statusMessage}
-                      </Typography>
-                    </Box>
-                  )}
-
-                  {env.deployment.image && (
-                    <>
-                      <Box display="flex" alignItems="center" mt={2}>
-                        <Typography variant="body2" color="textSecondary">
-                          Image
-                        </Typography>
-                      </Box>
-                      <Box
-                        display="flex"
-                        alignItems="center"
-                        className={classes.imageContainer}
-                      >
-                        <Typography
-                          variant="body2"
-                          color="textSecondary"
-                          style={{ wordBreak: 'break-all' }}
-                        >
-                          {env.deployment.image}
-                        </Typography>
-                      </Box>
-                    </>
-                  )}
-
-                  {env.deployment.status === 'success' &&
-                    env.endpoints.length > 0 && (
-                      <>
-                        <Box display="flex" alignItems="center" mt={2}>
-                          <Typography variant="body2" color="textSecondary">
-                            Endpoints
-                          </Typography>
-                        </Box>
-                        {env.endpoints.map((endpoint, index) => (
-                          <Box
-                            key={index}
-                            display="flex"
-                            alignItems="center"
-                            mt={index === 0 ? 0 : 1}
-                            sx={{ minWidth: 0, width: '100%' }}
-                          >
-                            <Box sx={{ flex: 1, minWidth: 0, mr: 1 }}>
-                              <a
-                                href={endpoint.url}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className={classes.endpointLink}
-                              >
-                                {endpoint.url}
-                              </a>
-                            </Box>
-                            <Box sx={{ flexShrink: 0 }}>
-                              <IconButton
-                                size="small"
-                                onClick={() => {
-                                  navigator.clipboard.writeText(endpoint.url);
-                                  // You could add a toast notification here
-                                }}
-                              >
-                                <FileCopyIcon fontSize="small" />
-                              </IconButton>
-                            </Box>
-                          </Box>
-                        ))}
-                      </>
-                    )}
-
-                  {/* Actions section - show if deployment is successful or suspended */}
-                  {((env.deployment.status === 'success' &&
-                    env.promotionTargets &&
-                    env.promotionTargets.length > 0) ||
-                    ((env.deployment.status === 'success' ||
-                      env.deployment.status === 'suspended') &&
-                      env.bindingName)) && (
-                    <Box mt={3}>
-                      {/* Multiple promotion targets - stack vertically */}
-                      {env.deployment.status === 'success' &&
-                        env.promotionTargets &&
-                        env.promotionTargets.length > 1 &&
-                        env.promotionTargets.map((target, index) => (
-                          <Box
-                            key={target.name}
-                            mb={
-                              index < env.promotionTargets!.length - 1
-                                ? 2
-                                : (env.deployment.status === 'success' ||
-                                    env.deployment.status === 'suspended') &&
-                                  env.bindingName
-                                ? 2
-                                : 0
-                            }
-                          >
-                            <Button
-                              variant="contained"
-                              color="primary"
-                              size="small"
-                              disabled={promotingTo === target.name}
-                              onClick={async () => {
-                                try {
-                                  setPromotingTo(target.name);
-                                  const result = await promoteToEnvironment(
-                                    entity,
-                                    discovery,
-                                    identityApi,
-                                    env.name.toLowerCase(), // source environment
-                                    target.name.toLowerCase(), // target environment
-                                  );
-
-                                  // Update environments state with fresh data from promotion result
-                                  setEnvironmentsData(result as Environment[]);
-
-                                  setNotification({
-                                    message: `Component promoted from ${env.name} to ${target.name}`,
-                                    type: 'success',
-                                  });
-
-                                  // Clear notification after 5 seconds
-                                  setTimeout(() => setNotification(null), 5000);
-                                } catch (err) {
-                                  setNotification({
-                                    message: `Error promoting: ${err}`,
-                                    type: 'error',
-                                  });
-
-                                  // Clear notification after 7 seconds for errors
-                                  setTimeout(() => setNotification(null), 7000);
-                                } finally {
-                                  setPromotingTo(null);
-                                }
-                              }}
-                            >
-                              {promotingTo === target.name
-                                ? 'Promoting...'
-                                : `Promote to ${target.name}`}
-                              {target.requiresApproval &&
-                                !promotingTo &&
-                                ' (Approval Required)'}
-                            </Button>
-                          </Box>
-                        ))}
-
-                      {/* Single promotion target and suspend button - show in same row */}
-                      {((env.deployment.status === 'success' &&
-                        env.promotionTargets &&
-                        env.promotionTargets.length === 1) ||
-                        ((env.deployment.status === 'success' ||
-                          env.deployment.status === 'suspended') &&
-                          env.bindingName)) && (
-                        <Box display="flex" flexWrap="wrap">
-                          {/* Single promotion button */}
-                          {env.deployment.status === 'success' &&
-                            env.promotionTargets &&
-                            env.promotionTargets.length === 1 && (
-                              <Button
-                                style={{ marginRight: '8px' }}
-                                variant="contained"
-                                color="primary"
-                                size="small"
-                                disabled={
-                                  promotingTo === env.promotionTargets[0].name
-                                }
-                                onClick={async () => {
-                                  try {
-                                    setPromotingTo(
-                                      env.promotionTargets![0].name,
-                                    );
-                                    const result = await promoteToEnvironment(
-                                      entity,
-                                      discovery,
-                                      identityApi,
-                                      env.name.toLowerCase(), // source environment
-                                      env.promotionTargets![0].name.toLowerCase(), // target environment
-                                    );
-
-                                    // Update environments state with fresh data from promotion result
-                                    setEnvironmentsData(
-                                      result as Environment[],
-                                    );
-
-                                    setNotification({
-                                      message: `Component promoted from ${
-                                        env.name
-                                      } to ${env.promotionTargets![0].name}`,
-                                      type: 'success',
-                                    });
-
-                                    // Clear notification after 5 seconds
-                                    setTimeout(
-                                      () => setNotification(null),
-                                      5000,
-                                    );
-                                  } catch (err) {
-                                    setNotification({
-                                      message: `Error promoting: ${err}`,
-                                      type: 'error',
-                                    });
-
-                                    // Clear notification after 7 seconds for errors
-                                    setTimeout(
-                                      () => setNotification(null),
-                                      7000,
-                                    );
-                                  } finally {
-                                    setPromotingTo(null);
-                                  }
-                                }}
-                              >
-                                {promotingTo === env.promotionTargets[0].name
-                                  ? 'Promoting...'
-                                  : 'Promote'}
-                                {env.promotionTargets[0].requiresApproval &&
-                                  !promotingTo &&
-                                  ' (Approval Required)'}
-                              </Button>
-                            )}
-
-                          {/* Suspend/Re-deploy button */}
-                          {(env.deployment.status === 'success' ||
-                            env.deployment.status === 'suspended') &&
-                            env.bindingName && (
-                              <Button
-                                variant="outlined"
-                                color={
-                                  env.deployment.status === 'suspended'
-                                    ? 'primary'
-                                    : 'default'
-                                }
-                                size="small"
-                                disabled={updatingBinding === env.name}
-                                onClick={async () => {
-                                  try {
-                                    setUpdatingBinding(env.name);
-                                    const newState =
-                                      env.deployment.status === 'suspended'
-                                        ? 'Active'
-                                        : 'Suspend';
-                                    await updateComponentBinding(
-                                      entity,
-                                      discovery,
-                                      identityApi,
-                                      env.bindingName!,
-                                      newState,
-                                    );
-
-                                    // Refresh the environments data
-                                    await fetchEnvironmentsData();
-
-                                    setNotification({
-                                      message: `Deployment ${
-                                        newState === 'Active'
-                                          ? 're-deployed'
-                                          : 'suspended'
-                                      } successfully`,
-                                      type: 'success',
-                                    });
-
-                                    // Clear notification after 5 seconds
-                                    setTimeout(
-                                      () => setNotification(null),
-                                      5000,
-                                    );
-                                  } catch (err) {
-                                    setNotification({
-                                      message: `Error updating deployment: ${err}`,
-                                      type: 'error',
-                                    });
-
-                                    // Clear notification after 7 seconds for errors
-                                    setTimeout(
-                                      () => setNotification(null),
-                                      7000,
-                                    );
-                                  } finally {
-                                    setUpdatingBinding(null);
-                                  }
-                                }}
-                              >
-                                {updatingBinding === env.name
-                                  ? 'Updating...'
-                                  : env.deployment.status === 'suspended'
-                                  ? 'Re-deploy'
-                                  : 'Suspend'}
-                              </Button>
-                            )}
-                        </Box>
-                      )}
-                    </Box>
-                  )}
-                </CardContent>
-              </Card>
-            </Grid>
-          ))}
-        </Grid>
-      </Content>
-    </Page>
+                )}
+              </CardContent>
+            </Box>
+          </Card>
+        </Box>
+        {environments.map(env => (
+          <EnvCard
+            key={env.name}
+            env={env}
+            entity={entity}
+            discovery={discovery}
+            identityApi={identityApi}
+            promotingTo={promotingTo}
+            setPromotingTo={setPromotingTo}
+            updatingBinding={updatingBinding}
+            setUpdatingBinding={setUpdatingBinding}
+            setEnvironmentsData={setEnvironmentsData}
+            setNotification={setNotification}
+            fetchEnvironmentsData={fetchEnvironmentsData}
+          />
+        ))}
+      </Box>
+    </Box>
+    //   </Content>
+    // </Page>
   );
 };
