+++ plugins/openchoreo/src/components/Environments/Workload/WorkloadEditor/index.ts
@@ -2,3 +2,4 @@ export { WorkloadEditor } from './WorkloadEditor';
 export { ContainerSection } from './ContainerSection';
 export { EndpointSection } from './EndpointSection';
 export { ConnectionSection } from './ConnectionSection';
+export { useWorkloadEditorStyles } from './styles';
